Fixed the ordering problem!  Woot!
There are now two field orders.  One for Member and Requester ctors, and
one for printing.  The _fields dict is now a dict of lists in order to
implement that.  The logic is all contained in ordered_field_names().

All methods that returned field names or values now take an 'order' arg
indicating which order they should return (default to ctor).

Squashed one or two places that still had the old order in a ctor.  That
took way too long to track down.

Added another test, but they really need work.

Rearranged the process form to make it a little more cohesive.