Add modify ability; break some long lines
Implement the ability to modify a record.  This ONLY updates the
database, and has no effect if the request was canceled.  Soon it will
be an error to request a cancelation along with either of the other two
actions, but for now cancelation takes precedence.

Also add some protection against malicious query strings, limiting the
fields to those defined in the member lib, and preventing malformed
request ids.