Implement registration; security fixes; i/f tweaks
Create register.py - holding the functions which govern registration
itself.  add_member_to_db() - which will just call out to the member
tool on trilug.org, and queue_for_shell(), which adds the appropriate
info to a queue file which can be consumed from the commandline by
someone with sufficient perms on the system.  Rename register to
request_membership to avoid naming confusion.

Move db file info into register.py.  Improves abstraction a touch.

Add cleanse_and_validate() to reg_scrub lib - runs the form through the
sanitizer as before, but then does some data validation on the form
vars, dropping any request that doesn't conform.

Add new libs, names to Makefile.

Overload __getitem__ and __setitem__ for a Member, to allow for easier
setting of fields.  It's a pretty thin abstraction, but it at least
allows us to dump the set_addr2 ugliness.

Add shell var fetchers to Member for use in populating shell requests in
the queue.

Add the request id to the process_registrations form (not as a var, just
as information for the person doing the registrations).

Add the candidate() method to the db - allows for the retrieval of a
single request from the database.