Schema tweak; validator change; error logging
Change the zipcode from an INT field to a TEXT - because it is.

Change the validation accordingly, as well as how it is handled in the
db code.

Add _print_err() to reg_scrub.py to print errors to stderr (shows up in
Apache's error log).  Probably going to pull this out into something
that's included in all the scripts.

Change _is_numeric() to _is_integer(), and likewise change the test
cases.